Move JsonReader to log4j-api
This parser is useful for LOG4J2-3415 and https://cwiki.apache.org/confluence/display/LOGGING/Properties+Enhancement
