+++ log4j-api/src/main/java/org/apache/logging/log4j/util/JsonReader.java
@@ -14,7 +14,7 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-package org.apache.logging.log4j.layout.template.json.util;
+package org.apache.logging.log4j.util;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
@@ -185,10 +185,24 @@ public final class JsonReader {
     }
 
     private void skipWhiteSpace() {
+        boolean inComment = false;
         do {
-            if (!Character.isWhitespace(readChar)) {
+            if (!inComment && readChar == '/') {
+                if (readChar() == '*') {
+                    inComment = true;
+                } else {
+                    unreadChar();
+                }
+            } else if (inComment && readChar == '*') {
+                if (readChar() == '/') {
+                    inComment = false;
+                } else {
+                    unreadChar();
+                }
+            } else if (!inComment && !Character.isWhitespace(readChar)) {
                 break;
             }
+
         } while (readChar() != CharacterIterator.DONE);
     }
